Better dependency management
Without this, KaMPIng does not work in conjuction with code which also
depends on cereal.

This requires us to bump to minimum required CMake version to
3.24 (shipped by all major distributions/easy to install).

This first tries to find a system installed cereal, and otherwise
fetches it via FetchContent and builds from source.

This also provides an option to completely disable serialization,
thereby getting rid of the dependency.

More dependency/CMake cleanup will be part of another PR.